+++ tf/env/staging/namespaces.tf
@@ -66,3 +66,11 @@ resource "kubernetes_resource_quota" "qs-jobs-podquota" {
     scopes = ["BestEffort"]
   }
 }
+
+resource "kubernetes_namespace" "tailscale-namespace" {
+  provider = kubernetes.wbaas-2
+
+  metadata {
+    name = "tailscale"
+  }
+}